FIX: Adjusted test so that initial `df` also has `txrx`
This is necessary, since `ComlinkChannel.__init__` generates
a `txrx` column in `ComlinkChannel.data`. Before the optimization
this seemed to have worked on a view of the initial Dataframe that
is passed to `__ini__` so that it also got a `txrx` column. Now,
since a new DataFrame is generated (because it is faster, for
whatever reason...) the initial Dataframe stays as it is (which
actually is the expected behavior).